Updates workflow header redesign based on feedback
